Rename updater suffix
