+++ pkg/gsm-secrets/secrets.go
@@ -8,9 +8,10 @@ import (
 	"github.com/openshift/ci-tools/pkg/group"
 )
 
-// ValidateSecretName validates if a secret name matches the allowed pattern
+// ValidateSecretName validates if a secret name matches the allowed pattern.
+// The secret name must be between 1 and 255 characters long (GCP limit).
 func ValidateSecretName(secretName string) bool {
-	return regexp.MustCompile(SecretNameRegex).MatchString(secretName)
+	return regexp.MustCompile(SecretNameRegex).MatchString(secretName) && len(secretName) <= 255
 }
 
 // ClassifySecret determines the type of secret based on its name
