+++ pkg/gsm-secrets/testdata/zz_fixture_bindings_TestGetDesiredState_one_secret_collection.yaml
@@ -13,7 +13,7 @@
   members:
   - group:team-alpha@redhat.com
   - group:team-beta@redhat.com
-  - serviceAccount:collection-alpha-updater@test-project.iam.gserviceaccount.com
+  - serviceAccount:collection-alpha-sa@test-project.iam.gserviceaccount.com
   role: projects/test-project/roles/openshift_ci_secrets_viewer
 - condition:
     description: 'Managed by test platform: Create, update, and delete access to secrets
@@ -24,5 +24,5 @@
   members:
   - group:team-alpha@redhat.com
   - group:team-beta@redhat.com
-  - serviceAccount:collection-alpha-updater@test-project.iam.gserviceaccount.com
+  - serviceAccount:collection-alpha-sa@test-project.iam.gserviceaccount.com
   role: projects/test-project/roles/openshift_ci_secrets_updater
