+++ pkg/gsm-secrets/testdata/zz_fixture_sa_TestGetDesiredState_basic_config.yaml
@@ -1,9 +1,9 @@
 - Collection: collection-alpha
-  DisplayName: collection-alpha-updater
-  Email: collection-alpha-updater@test-project.iam.gserviceaccount.com
+  DisplayName: collection-alpha-sa
+  Email: collection-alpha-sa@test-project.iam.gserviceaccount.com
 - Collection: collection-beta
-  DisplayName: collection-beta-updater
-  Email: collection-beta-updater@test-project.iam.gserviceaccount.com
+  DisplayName: collection-beta-sa
+  Email: collection-beta-sa@test-project.iam.gserviceaccount.com
 - Collection: shared-collection
-  DisplayName: shared-collection-updater
-  Email: shared-collection-updater@test-project.iam.gserviceaccount.com
+  DisplayName: shared-collection-sa
+  Email: shared-collection-sa@test-project.iam.gserviceaccount.com
