+++ pkg/gsm-secrets/testdata/zz_fixture_sa_TestGetDesiredState_complex_config.yaml
@@ -1,9 +1,9 @@
 - Collection: alpha-secrets
-  DisplayName: alpha-secrets-updater
-  Email: alpha-secrets-updater@test-project.iam.gserviceaccount.com
+  DisplayName: alpha-secrets-sa
+  Email: alpha-secrets-sa@test-project.iam.gserviceaccount.com
 - Collection: beta-delta-secrets
-  DisplayName: beta-delta-secrets-updater
-  Email: beta-delta-secrets-updater@test-project.iam.gserviceaccount.com
+  DisplayName: beta-delta-secrets-sa
+  Email: beta-delta-secrets-sa@test-project.iam.gserviceaccount.com
 - Collection: epsilon-secrets
-  DisplayName: epsilon-secrets-updater
-  Email: epsilon-secrets-updater@test-project.iam.gserviceaccount.com
+  DisplayName: epsilon-secrets-sa
+  Email: epsilon-secrets-sa@test-project.iam.gserviceaccount.com
