+++ pkg/gsm-secrets/testdata/zz_fixture_sa_TestGetDesiredState_one_secret_collection.yaml
@@ -1,3 +1,3 @@
 - Collection: collection-alpha
-  DisplayName: collection-alpha-updater
-  Email: collection-alpha-updater@test-project.iam.gserviceaccount.com
+  DisplayName: collection-alpha-sa
+  Email: collection-alpha-sa@test-project.iam.gserviceaccount.com
